Add files via upload
In this file, I set out to show how small differences in means and variances throw the incorrect understanding of NHST that is sometimes advertised in the wild. Since there is no true zero effect and NHST makes no statement about the size of the effect, the intepretation of p-values is often wrong.